T-AIA-911-STG_1/services/nlp-service/app/services/classifieur.py
from __future__ import annotations

from sklearn.pipeline import Pipeline
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression


def build_is_valid_classifier() -> Pipeline:
    return Pipeline([
        ("tfidf", TfidfVectorizer(
            analyzer="char_wb",
            ngram_range=(3, 5),
            min_df=2
        )),
        ("clf", LogisticRegression(max_iter=2000))
    ])
