T-AIA-911-STG_1/services/nlp-service/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
httpx==0.27.2
rapidfuzz==3.14.3
codecarbon==3.2.0
psutil==7.2.1
py-cpuinfo==9.0.0
nvidia-ml-py==13.590.44
prometheus_client==0.23.1
pydantic==2.12.5
pydantic-core==2.41.5
scikit-learn
python-dotenv
pandas
sklearn-crfsuite
joblib
